moved /ban command into it's own file
